Move sysusers.d/sysctl.d/binfmt.d/modules-load.d back to /usr
In Debian, late mounting of /usr is no longer supported, so it is safe
to install those files in /usr.
We want those facilities in /usr, not /, as this will make an eventual
switch to a merged-usr setup easier.

Closes: #971282

Gbp-Pq: Topic debian
Gbp-Pq: Name Move-sysusers.d-sysctl.d-binfmt.d-modules-load.d-back-to-.patch