test-execute: skip tests that are broken without unprivileged userns
Origin: upstream, https://github.com/systemd/systemd/commit/d0c6136f5130ee5baa561e3af5a3b53719588dcf

With newer versions of AppArmor, unprivileged user namespace creation
may be restricted by default, in which case user manager instances will
not be able to apply PrivateUsers=yes (or the settings which require it).
Additionally, if a kernel has the kernel.unprivileged_userns_clone
sysctl patch, and that sysctl is 0, then unprivileged userns creation
will always fail.

If a test unit is going to be run in a user manager, and that unit
requires PrivateUsers=yes (explicitly or implicitly), then skip it if
we do not have user namespace privileges.

Gbp-Pq: Name test-execute-skip-tests-that-are-broken-without-unprivile.patch